Move fixture to duthost_utils.py
